Switch to Maven source-root convention
