+++ RandomJSONRPC/src/main/java/org/random/api/JsonObjectInputCallable.java
