+++ RandomJSONRPC/src/main/java/org/random/api/exception/RandomOrgBadHTTPResponseException.java
