+++ RandomJSONRPC/src/main/java/org/random/api/exception/RandomOrgInsufficientBitsError.java
