+++ RandomJSONRPC/src/main/java/org/random/api/exception/RandomOrgInsufficientRequestsError.java
