+++ RandomJSONRPC/src/main/java/org/random/api/exception/RandomOrgJSONRPCError.java
