+++ RandomJSONRPC/src/main/java/org/random/api/exception/RandomOrgRANDOMORGError.java
