+++ RandomJSONRPC/src/main/java/org/random/api/exception/RandomOrgSendTimeoutException.java
